refactor: overhaul personal portfolio with major updates
