+++ src/app/_components/common/Fragments/TechListCard.tsx
