+++ src/app/_components/containers/loaders/projectLoader.tsx
@@ -1,0 +1,7 @@
+export default function ProjectLoader() {
+  return (
+    <>
+      <div>Ini Adalah Project Loader</div>
+    </>
+  );
+}
